node: actions: executeWithdrawal: assume fees paid in withdrawal API (#255)
In this PR, we remove the `payFees` invocation in `executeWithdrawal`. Fee payment tasks are expected to be enqueued by the relayer as of https://github.com/renegade-fi/renegade/pull/1132.

The `withdraw` WASM util already replicates the effects of fee payments on the local wallet, so no changes are necessary there.

### Testing
Testing will be done downstream of integration in the frontend